+++ src/main/java/_4/TourismContest/baseball/dto/BaseballScheduleDTO.java
@@ -1,12 +1,10 @@
-package _4.TourismContest.crawler.dto;
+package _4.TourismContest.baseball.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
