+++ src/main/java/_4/TourismContest/baseball/repository/BaseballRepository.java
@@ -1,6 +1,6 @@
-package _4.TourismContest.crawler.repository;
+package _4.TourismContest.baseball.repository;
 
-import _4.TourismContest.crawler.domain.Baseball;
+import _4.TourismContest.baseball.domain.Baseball;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
